requirements.txt
bcrypt
fastapi
httpx
passlib
python-dotenv
python-jose
python-multipart
uvicorn
